# expression: implement vectorized evaluation for `builtinLikeSig`