--- a/expression/builtin_like_vec.go
+++ b/expression/builtin_like_vec.go
@@ -19,14 +19,58 @@ import (
 	"github.com/pingcap/errors"
 	"github.com/pingcap/tidb/types"
 	"github.com/pingcap/tidb/util/chunk"
+	"github.com/pingcap/tidb/util/stringutil"
 )
 
 func (b *builtinLikeSig) vectorized() bool {
-	return false
+	return true
 }
 
 func (b *builtinLikeSig) vecEvalInt(input *chunk.Chunk, result *chunk.Column) error {
-	return errors.Errorf("not implemented")
+	n := input.NumRows()
+	bufVal, err := b.bufAllocator.get(types.ETString, n)
+	if err != nil {
+		return err
+	}
+	defer b.bufAllocator.put(bufVal)
+	if err = b.args[0].VecEvalString(b.ctx, input, bufVal); err != nil {
+		return err
+	}
+
+	bufPattern, err := b.bufAllocator.get(types.ETString, n)
+	if err != nil {
+		return err
+	}
+	defer b.bufAllocator.put(bufPattern)
+	if err = b.args[1].VecEvalString(b.ctx, input, bufPattern); err != nil {
+		return err
+	}
+
+	bufEscape, err := b.bufAllocator.get(types.ETInt, n)
+	if err != nil {
+		return err
+	}
+	defer b.bufAllocator.put(bufEscape)
+	if err = b.args[2].VecEvalInt(b.ctx, input, bufEscape); err != nil {
+		return err
+	}
+	escapes := bufEscape.Int64s()
+
+	result.ResizeInt64(n, false)
+	result.MergeNulls(bufVal, bufPattern, bufEscape)
+	i64s := result.Int64s()
+	for i := 0; i < n; i++ {
+		if result.IsNull(i) {
+			continue
+		}
+
+		escape := byte(escapes[i])
+		patChars, patTypes := stringutil.CompilePattern(bufPattern.GetString(i), escape)
+		match := stringutil.DoMatch(bufVal.GetString(i), patChars, patTypes)
+		i64s[i] = boolToInt64(match)
+	}
+
+	return nil
 }
 
 func (b *builtinRegexpBinarySig) vectorized() bool {